About PyGoL
As part of the Programming Module of the Interdisciplinary Bioscience Doctoral Training Program, I have developed PyGoL, a python framework for a Game of Life-like cellular automaton, for modelling populations of organisms in an environment with renewable but limited resources available. This framework is an easy-to-use open-source tool, user-friendly and computationally efficient, without requiring installation of special python modules. It is a very versatile framework that allows the user to define a number of variables for specific population studies and can be easily adapted for other applications. The framework returns the output to the user in several formats, allowing for both fast interpretation and further analysis of results. Here I document the development and workflow of PyGoL, and give insights into the further developments and applications for this framework.